change scoreid to bigint

+++ database/objects.py
@@ -156,7 +156,7 @@ class DBScore(Base):
     
     __tablename__ = 'scores'
     
-    id = Column('id', Integer, primary_key=True)
+    id = Column('id', BigInteger, primary_key=True)
     user_id = Column('user_id', Integer)
     server = Column('server', String, primary_key=True)
     beatmap_id = Column('beatmap_id', Integer, ForeignKey('beatmaps.id'))
